added clear function
