+++ src/index.html
@@ -43,7 +43,9 @@
         <div><img class="arrow" src="/asset/resource/R.png" alt="arrow" /></div>
       </div>
       <ul id="todolist" class="list-unstyled"></ul>
-      <div class="text-center mb-3">Clear all completed</div>
+      <div class="text-center mb-3">
+        <button id="clear">Clear all completed</button>
+      </div>
     </div>
     <script type="module" src="./index.js"></script>
     <script type="module" src="/dist/main.js"></script>
