+++ webpack.config.js
@@ -10,6 +10,11 @@ module.exports = {
         test: /\.css$/i,
         use: ['style-loader', 'css-loader'],
       },
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: 'babel-loader'
+      }
     ],
   },
   entry: './src/index.js',
